Move Layout to top-level and move tutorials to their own route
