+++ ui/src/app/tutorials/page.tsx
@@ -13,14 +13,15 @@
 //limitations under the License.
 
 //SPDX-License-Identifier: Apache-2.0
-import { type PropsWithChildren } from "react";
+import { Container } from "~/components/Container";
+import { HomepageCards } from "~/components/HomepageCards";
+import { HomepageHero } from "~/components/HomepageHero";
 
-import { env } from "~/env.mjs";
-
-export const metadata = {
-  title: `${env.NEXT_PUBLIC_APP_NAME} Resources`,
-};
-
-export default function ResourcesLayout({ children }: PropsWithChildren) {
-  return children;
+export default function Dashboard() {
+  return (
+    <Container>
+      <HomepageHero />
+      <HomepageCards />
+    </Container>
+  );
 }
